+++ CHANGELOG.md
@@ -29,3 +29,4 @@ Put the changelog BELOW the dashes. ANYTHING ABOVE IS IGNORED.
   - The Item pickup animation now plays when Tumbleweed picks up an Item.
   - Tumbleweed's visual movement should now be slightly less clunky.
   - Revised the models of the Tumbleweed Stem's first two stages to be more consistent with Vanilla's style.
+- Placing Tumbleweed now spawns the Tumbleweed Entity.